move code from OrderItemController to OrderItemService
