+++ website/src/main/java/ru/teamtwo/website/controller/user/OrderController.java
@@ -11,7 +11,7 @@ import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.ResponseBody;
 import org.springframework.web.bind.annotation.RestController;
 import ru.teamtwo.core.dtos.user.OrderDto;
-import ru.teamtwo.website.service.OrderService;
+import ru.teamtwo.website.service.user.OrderService;
 
 @Slf4j
 @RestController
