+++ website/src/main/java/ru/teamtwo/website/service/user/CartItemService.java
@@ -1,4 +1,4 @@
-package ru.teamtwo.website.service;
+package ru.teamtwo.website.service.user;
 
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Autowired;
