+++ website/src/main/java/ru/teamtwo/website/service/user/CustomerService.java
@@ -1,4 +1,4 @@
-package ru.teamtwo.website.service;
+package ru.teamtwo.website.service.user;
 
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -7,7 +7,6 @@ import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Service;
 import ru.teamtwo.core.dtos.user.CustomerDto;
 import ru.teamtwo.core.models.user.Customer;
-import ru.teamtwo.website.repository.user.CartItemRepository;
 import ru.teamtwo.website.repository.user.CustomerRepository;
 
 @Slf4j
